Simple moving of functions from cli.py to utilities files
