+++ src/hdx_cli_toolkit/utilities.py
@@ -3,11 +3,14 @@
 
 import csv
 import dataclasses
+import datetime
 import json
 import os
 
 from collections.abc import Callable
 from typing import Any
+
+import click
 
 
 def write_dictionary(
@@ -263,3 +266,19 @@ def read_attributes(dataset_name: str, attributes_filepath: str) -> dict:
         attributes = {}
 
     return attributes
+
+
+def print_banner(action: str):
+    """Simple function to output a banner to console, uses click's secho command but not colour
+    because the underlying colorama does not output correctly to git-bash terminals.
+
+    Arguments:
+        action {str} -- _description_
+    """
+    title = f"HDX CLI toolkit - {action}"
+    timestamp = f"Invoked at: {datetime.datetime.now().isoformat()}"
+    width = max(len(title), len(timestamp))
+    click.secho((width + 4) * "*", bold=True)
+    click.secho(f"* {title:<{width}} *", bold=True)
+    click.secho(f"* {timestamp:<{width}} *", bold=True)
+    click.secho((width + 4) * "*", bold=True)
